Refactor control headers.
